requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
DAWG-Python==0.7.2
Django==5.0.3
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
docopt==0.6.2
idna==3.7
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.5
murmurhash==1.0.10
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
preshed==3.0.9
pydantic==2.7.0
pydantic_core==2.18.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.4.16
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
thinc==8.2.3
threadpoolctl==3.4.0
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
